refactor: topic and consume produce
